rename package to jmxmetricextension
