+++ Makefile
@@ -17,7 +17,7 @@ OTEL_VERSION=master
 # Modules to run integration tests on.
 # XXX: Find a way to automatically populate this. Too slow to run across all modules when there are just a few.
 INTEGRATION_TEST_MODULES := \
-	extension/jmxmetricsextension/ \
+	extension/jmxmetricextension/ \
 	receiver/dockerstatsreceiver \
 	receiver/redisreceiver \
 	internal/common
