+++ extension/jmxmetricextension/testdata/Dockerfile.cassandra
