+++ extension/jmxmetricextension/testdata/script.groovy
